# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
export interface Ok {
  ok: boolean;
}

export type AccountTier = typeof AccountTier[keyof typeof AccountTier];


export const AccountTier = {
  operator: 'operator',
  tester: 'tester',
} as const;

export interface Account {
  id: string;
  email: string;
  tier: AccountTier;
}

export interface EmailInput {
  email: string;
}

export interface TokenInput {
  token: string;
}

export type SourceRole = typeof SourceRole[keyof typeof SourceRole];


export const SourceRole = {
  camera_a: 'camera_a',
  camera_b: 'camera_b',
  audio: 'audio',
} as const;

export interface Source {
  id: string;
  role: SourceRole;
  ingestState: string;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  fps?: number | null;
}

export interface Shoot {
  id: string;
  name: string;
  /** @nullable */
  shotOn?: string | null;
  /** @nullable */
  notes?: string | null;
  sources: Source[];
}

export interface ShootInput {
  /** @minLength 1 */
  name: string;
  shotOn?: string;
  notes?: string;
}

export type UploadInputRole = typeof UploadInputRole[keyof typeof UploadInputRole];


export const UploadInputRole = {
  camera_a: 'camera_a',
  camera_b: 'camera_b',
  audio: 'audio',
} as const;

export interface UploadInput {
  shootId: string;
  role: UploadInputRole;
  fileName: string;
  contentType: string;
}

export interface Upload {
  id: string;
  uploadId: string;
}

export interface PartInput {
  uploadId: string;
  /**
     * @minimum 1
     * @maximum 10000
     */
  partNumber: number;
}

export type PartsInputPartsItem = {
  partNumber: number;
  etag: string;
};

export interface PartsInput {
  uploadId: string;
  parts: PartsInputPartsItem[];
}

export interface SignedUrl {
  url: string;
}

export interface Question {
  id: string;
  feature: string;
  key: string;
  prompt: string;
  position: number;
}

export type FeedbackInputState = { [key: string]: unknown };

export interface FeedbackInput {
  questionKey: string;
  answer: string;
  page: string;
  state: FeedbackInputState;
}

export type FeedbackState = { [key: string]: unknown };

export interface Feedback {
  id: string;
  userId: string;
  questionKey: string;
  answer: string;
  page: string;
  buildVersion: string;
  state: FeedbackState;
  createdAt: string;
}

export type ListQuestionsParams = {
feature: string;
};