# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { AccountTier } from './accountTier';

export interface Account {
  id: string;
  email: string;
  tier: AccountTier;
}