# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackState } from './feedbackState';

export interface Feedback {
  id: string;
  userId: string;
  questionKey: string;
  answer: string;
  page: string;
  buildVersion: string;
  state: FeedbackState;
  createdAt: string;
}