# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputState } from './feedbackInputState';

export interface FeedbackInput {
  questionKey: string;
  answer: string;
  page: string;
  state: FeedbackInputState;
}