# lib/api-zod/src/generated/types/partInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface PartInput {
  uploadId: string;
  /**
     * @minimum 1
     * @maximum 10000
     */
  partNumber: number;
}