# lib/api-zod/src/generated/types/partsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { PartsInputPartsItem } from './partsInputPartsItem';

export interface PartsInput {
  uploadId: string;
  parts: PartsInputPartsItem[];
}