# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: string;
  feature: string;
  key: string;
  prompt: string;
  position: number;
}