# lib/api-zod/src/generated/types/shoot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { Source } from './source';

export interface Shoot {
  id: string;
  name: string;
  /** @nullable */
  shotOn?: string | null;
  /** @nullable */
  notes?: string | null;
  sources: Source[];
}