# lib/api-zod/src/generated/types/shootInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface ShootInput {
  /** @minLength 1 */
  name: string;
  shotOn?: string;
  notes?: string;
}