# lib/api-zod/src/generated/types/source.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { SourceRole } from './sourceRole';

export interface Source {
  id: string;
  role: SourceRole;
  ingestState: string;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  fps?: number | null;
}